API

ExprTools.combinedef

ExprTools.signature

ExprTools.splitdef

ExprTools.splitdef — Function

splitdef(ex::Expr; throw::Bool=true) -> Union{Dict{Symbol,Any}, Nothing}

Split a function definition expression into its various components including:

:head: Expression head of the function definition (:function, :(=), :(->))

:name: Name of the function (not present for anonymous functions)

:params: Parametric types defined on constructors

:args: Positional arguments of the function

:kwargs: Keyword arguments of the function

:rtype: Return type of the function

:whereparams: Where parameters

:body: Function body (not present for empty functions)

All components listed may not be present in the returned dictionary with the exception of :head which will always be present.

If the provided expression is not a function then an exception will be raised when throw=true. Use throw=false avoid raising an exception and return nothing instead.

See also: combinedef

ExprTools.combinedef — Function

combinedef(def::Dict{Symbol,Any}) -> Expr

Create a function definition expression from various components. Typically used to construct a function using the result of splitdef.

If def[:head] is not provided it will default to :function.

For more details see the documentation on splitdef.

ExprTools.signature — Function

signature(m::Method) -> Dict{Symbol,Any}

Finds the expression for a method's signature as broken up into its various components including:

:name: Name of the function

:params: Parametric types defined on constructors

:args: Positional arguments of the function

:whereparams: Where parameters

All components listed above may not be present in the returned dictionary if they are not in the function definition.

Limited support for:

:kwargs: Keyword arguments of the function. Only the names will be included, not the default values or type constraints.

Unsupported:

:rtype: Return type of the function

:body: Function body0

:head: Expression head of the function definition (:function, :(=), :(->))

For more complete coverage, consider using splitdef with CodeTracking.definition.

The dictionary of components returned by signature match those returned by splitdef and include all that are required by combinedef, except for the :body component.